moved search bar implementation code from layout to SearchBar cmpnt
