+++ layouts/default.vue
@@ -49,30 +49,8 @@
           </span>
         </button>
 
-        <!-- SEARCH BAR -->
-        <div class="relative">
-          <div
-            class="absolute inset-y-0 right-0 flex cursor-pointer items-center pr-2"
-          >
-            <Icon
-              name="majesticons:search-line"
-              class="size-8 rounded-full bg-red-900/25 p-1 text-white hover:bg-red-900/50"
-              aria-hidden="true"
-              @click="doSearch(searchText)"
-            />
-          </div>
-          <input
-            v-model="searchText"
-            class="w-full bg-red-700 p-4 placeholder:font-semibold placeholder:text-white/80 focus:border-0 focus:bg-red-800 focus:outline-none dark:bg-red-800 dark:focus:bg-red-600"
-            placeholder="Search Open5e"
-            @keyup.enter="doSearch(searchText)"
-          />
-        </div>
-
-        <!-- Sidebar Navigation -->
+        <SearchBar />
         <NavMenu />
-
-        <!-- Report Issue UI -->
         <ModalReportIssue />
 
         <!-- Patron Banner -->
@@ -135,7 +113,6 @@
 </template>
 
 <script setup lang="ts">
-import { useRoute } from 'nuxt/app';
 import { computed } from 'vue';
 import EncounterBuilder from '~/components/EncounterBuilder.vue';
 import EncounterBuilderSummary from '~/components/EncounterBuilderSummary.vue';
@@ -151,13 +128,7 @@ useHead({ title: title });
 
 const showSidebar = ref(false);
 const isEncounterVisible = ref(false);
-const route = useRoute();
-watch(
-  () => route.path,
-  () => (showSidebar.value = false),
-);
-
-const searchText = ref(route.query.text);
+watch(() => (showSidebar.value = false));
 
 const showModal = ref(false);
 const { sources } = useSourcesList();
@@ -169,13 +140,6 @@ const { data: documents } = useDocuments({ fields: 'none' });
 const no_available_sources = computed(() => documents.value?.length ?? 0);
 
 const isLoadingData = useIsFetching();
-
-const router = useRouter();
-
-function doSearch(searchText) {
-  router.push({ name: 'search', query: { text: searchText } });
-  showSidebar.value = false;
-}
 
 const toggleSidebar = () => (showSidebar.value = !showSidebar.value);
 const hideSidebar = () => (showSidebar.value = false);
